+++ ChangeLog.md
@@ -5,6 +5,7 @@
 * Résistance à la magie du barbare
 
 ### Autres améliorations
+* Ajout du bonus de couvert
 * Attaques par défaut dans la liste d'actions, possibilité de se passer complètement des attributs pour les attaques (actif quand la version 3.3 de la fiche sera là.
 
 ### Corrections de bugs